Part 4 done

+++ part4/controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogsRouter = require("express").Router()
-
+const jwt = require("jsonwebtoken")
+const Blog = require("../models/blog")
 const Blogs = require("../models/blog")
 const User = require("../models/user")
 
@@ -22,12 +23,16 @@ blogsRouter.get("/:id", async (request, response) => {
 })
 blogsRouter.post("/", async (request, response) => {
   const body = request.body 
-  const user = await User.findById(body.userId)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({error: "token missing or invalid"})
+  }
+  const user = await User.findById(decodedToken.id)
   const blog = new Blogs({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes,
     user: user._id
   }) 
   const savedBlog = await blog.save()
@@ -47,9 +52,24 @@ blogsRouter.put("/:id", async (request, response) => {
   const updatedBlog = await Blogs.findByIdAndUpdate(request.params.id, blog, { new: true })
   response.json(updatedBlog) 
 }) 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blogs.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete("/:id", async (request, response, next) => {
+  const token = request.token
+  try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({error: "Token is invalid or missing"})
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (user._id.toString() === blog.user.toString()) {
+      await Blogs.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({error: "User is not permitted to delete the blog"}) 
+    }
+  } catch(error) {
+    next()
+  }
 })
 module.exports = blogsRouter
